requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
frozenlist==1.3.3
gunicorn==21.2.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
openai==0.27.8
packaging==23.1
path==16.6.0
pathlib==1.0.1
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.0.1
pydantic_core==2.0.2
pydub==0.25.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
regex==2023.6.3
requests==2.31.0
SQLAlchemy==2.0.17
srt==3.5.3
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.26.7
vosk==0.3.44
Wave==0.0.2
websockets==11.0.3
yarl==1.9.2
